# apps/managed-agent-api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Managed Agent API
  version: 0.1.0
  summary: Public frontend-facing HTTP and SSE contract for the managed-agent platform.
  description: >
    This specification defines the current browser-facing boundary exposed by
    `apps/managed-agent-api`. It is intentionally scoped to the Web UI and
    other first-party clients. Internal worker transport is excluded.
servers:
  - url: http://127.0.0.1:4173
    description: Local development server
...
paths:
...
  /me/llm-providers:
    get:
      tags: [LLM Providers]
      summary: List provider configs owned by the current user
      operationId: listUserLlmProviders
      security:
        - SessionCookieAuth: []
      responses:
        "200":
          description: User-owned provider configs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LlmProviderConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [LLM Providers]
      summary: Create one user-owned provider config
      operationId: createUserLlmProvider
      security:
        - SessionCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertLlmProviderConfigRequest"
      responses:
        "200":
          description: Provider config persisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmProviderConfig"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/llm-providers/{providerConfigId}:
    patch:
      tags: [LLM Providers]
      summary: Update one user-owned provider config
      operationId: updateUserLlmProvider
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/ProviderConfigIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchLlmProviderConfigRequest"
      responses:
        "200":
          description: Provider config updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmProviderConfig"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [LLM Providers]
      summary: Delete one user-owned provider config
      operationId: deleteUserLlmProvider
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/ProviderConfigIdPath"
      responses:
        "204":
          description: Provider config deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /me/llm-providers/{providerConfigId}/validate:
    post:
      tags: [LLM Providers]
      summary: Validate one user-owned provider config
      operationId: validateUserLlmProvider
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/ProviderConfigIdPath"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmProviderValidationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /me/llm-providers/{providerConfigId}/oauth/start:
    post:
      tags: [LLM Providers]
      summary: Start an OAuth browser flow for one user-owned provider config
      operationId: startUserLlmProviderOauthFlow
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/ProviderConfigIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LlmProviderOAuthStartRequest"
      responses:
        "200":
          description: OAuth flow started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmProviderOAuthFlow"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /me/llm-providers/{providerConfigId}/oauth/flows/{flowId}:
    get:
      tags: [LLM Providers]
      summary: Get one in-flight OAuth browser flow
      operationId: getUserLlmProviderOauthFlow
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/ProviderConfigIdPath"
        - $ref: "#/components/parameters/OAuthFlowIdPath"
      responses:
        "200":
          description: OAuth flow status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmProviderOAuthFlow"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /me/llm-providers/{providerConfigId}/oauth-account:
    delete:
      tags: [LLM Providers]
      summary: Remove the stored OAuth credential from one user-owned provider config
      operationId: deleteUserLlmProviderOauthAccount
      security:
        - SessionCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/ProviderConfigIdPath"
      responses:
        "200":
          description: Provider config updated after credential removal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmProviderConfig"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /oauth/llm-provider-flows/openai-codex/callback:
    get:
      tags: [LLM Providers]
      summary: Browser callback target for OpenAI Codex OAuth flows
      operationId: completeOpenAICodexOauthCallback
      parameters:
        - in: query
          name: state
          required: false
          schema:
            type: string
        - in: query
          name: code
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OAuth callback completed
          content:
            text/html:
              schema:
                type: string
        "400":
          description: Callback query is missing required parameters
          content:
            text/html:
              schema:
                type: string
        "404":
          description: OAuth flow was not found
          content:
            text/html:
              schema:
                type: string
        "500":
          description: OAuth callback failed
          content:
            text/html:
              schema:
                type: string
...
components:
  securitySchemes:
    SessionCookieAuth:
      type: apiKey
      in: cookie
      name: managed_agent_session
  parameters:
...
    ProviderConfigIdPath:
      name: providerConfigId
      in: path
      required: true
      schema:
        type: string
    OAuthFlowIdPath:
      name: flowId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: bad_request
              message: Invalid request parameters
    Unauthorized:
      description: Request requires an authenticated login session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: unauthorized
              message: login session required
...
    NotFound:
      description: Requested resource does not exist in the current user scope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: not_found
              message: resource not found
...
  schemas:
...
    LlmProviderModel:
      type: object
      required: [modelId, displayName, supportsReasoning]
      properties:
        modelId:
          type: string
        displayName:
          type: string
        supportsReasoning:
          type: boolean
    OAuthCredential:
      type: object
      required: [access, refresh, expires]
      properties:
        access:
          type: string
        refresh:
          type: string
        expires:
          type: integer
          minimum: 0
        accountId:
          type: string
        enterpriseUrl:
          type: string
...
    LlmProviderConfig:
      type: object
      required:
        [
          providerConfigId,
          providerType,
          displayName,
          authMode,
          headers,
          availableModels,
          defaultModelId,
          defaultThinkingLevel,
          enabled,
          hasStoredCredential
        ]
      properties:
        providerConfigId:
          type: string
        providerType:
          type: string
        displayName:
          type: string
        authMode:
          type: string
          enum: [api_key, oauth, none]
        baseUrl:
          type: string
        headers:
          type: object
          additionalProperties:
            type: string
        availableModels:
          type: array
          items:
            $ref: "#/components/schemas/LlmProviderModel"
        defaultModelId:
          type: string
        fastModelId:
          type: string
        balancedModelId:
          type: string
        strongModelId:
          type: string
        defaultThinkingLevel:
          type: string
        enabled:
          type: boolean
        hasStoredCredential:
          type: boolean
    UpsertLlmProviderConfigRequest:
      type: object
      required: [providerType]
      properties:
        providerType:
          type: string
          minLength: 1
        displayName:
          type: string
          minLength: 1
        baseUrl:
          type: string
          minLength: 1
        headers:
          type: object
          additionalProperties:
            type: string
        availableModels:
          type: array
          minItems: 1
          items:
            type: string
            minLength: 1
        defaultModelId:
          type: string
          minLength: 1
        fastModelId:
          type: string
          minLength: 1
        balancedModelId:
          type: string
          minLength: 1
        strongModelId:
          type: string
          minLength: 1
        defaultThinkingLevel:
          type: string
          minLength: 1
        enabled:
          type: boolean
        apiKey:
          type: string
          minLength: 1
        oauthCredential:
          $ref: "#/components/schemas/OAuthCredential"
    PatchLlmProviderConfigRequest:
      type: object
      properties:
        providerType:
          type: string
          minLength: 1
        displayName:
          type: string
          minLength: 1
        baseUrl:
          type: string
          minLength: 1
        headers:
          type: object
          additionalProperties:
            type: string
        availableModels:
          type: array
          minItems: 1
          items:
            type: string
            minLength: 1
        defaultModelId:
          type: string
          minLength: 1
        fastModelId:
          type: string
          minLength: 1
        balancedModelId:
          type: string
          minLength: 1
        strongModelId:
          type: string
          minLength: 1
        defaultThinkingLevel:
          type: string
          minLength: 1
        enabled:
          type: boolean
        apiKey:
          type: string
          minLength: 1
        oauthCredential:
          $ref: "#/components/schemas/OAuthCredential"
    LlmProviderValidationResponse:
      type: object
      required: [providerConfigId, valid, errors]
      properties:
        providerConfigId:
          type: string
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: string
    LlmProviderOAuthStartRequest:
      type: object
      properties:
        enterpriseUrl:
          type: string
          minLength: 1
    LlmProviderOAuthFlow:
      type: object
      required: [flowId, status, expiresAt, usesManagedCallback]
      properties:
        flowId:
          type: string
        status:
          type: string
          enum: [pending, completed, failed]
        authUrl:
          type: string
        instructions:
          type: string
        error:
          type: string
        expiresAt:
          type: integer
        completedAt:
          type: string
        usesManagedCallback:
          type: boolean
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...